+++ init.go
@@ -7,7 +7,6 @@ import (
 	"os"
 
 	"github.com/Songmu/prompter"
-	"github.com/aws/aws-sdk-go/service/codedeploy"
 	"github.com/aws/aws-sdk-go/service/ecs"
 	"github.com/pkg/errors"
 	"gopkg.in/yaml.v2"
@@ -102,17 +101,4 @@ func (d *App) saveFile(path string, b []byte, mode os.FileMode) error {
 		}
 	}
 	return ioutil.WriteFile(path, b, mode)
-}
-
-func (d *App) findDeployment(sv *ecs.Service) (*codedeploy.DeploymentInfo, error) {
-	if len(sv.TaskSets) == 0 {
-		return nil, errors.New("taskSet is not found in service")
-	}
-	dp, err := d.codedeploy.GetDeployment(&codedeploy.GetDeploymentInput{
-		DeploymentId: sv.TaskSets[0].ExternalId,
-	})
-	if err != nil {
-		return nil, err
-	}
-	return dp.DeploymentInfo, nil
 }
